refactor: rename fileserver -> cinema
